Add `IdentifiedContainer`, new `DiffableListsOfState` UIKit example
While the main TCA repo is iOS 13.0+, thankfully RAS-TCA has lowered
that requirement which allows the less "fortunate" of us to use it in
our projects 🙏.

Nonetheless, there are still some "goodies" in RAS-TCA that are iOS 13+,
one of them being the `Reducer.forEach` primitive that uses
`IdentifiedArray`. As `IdentifiedArray` builds upon `Identifiable` for
some of its conformances and extensions which is iOS 13+, it restricts
this whole set of functionality to iOS 13+.

By creating a new `IdentifiableContainer` protocol that is not
dependent on `Identifiable` and using that on the mentioned
`Reducer.forEach`, we can remove this restriction. Conformance to
`IdentifiedArray` is added so everything is backwards compatible, and
anyone targeting lower versions than iOS 13 can implement their own
container that conforms to that protocol.

Some additional `Store.forEach` helpers were also created to easily
create scoped stores for a set of child states.

A new `Store.childStore` helper was created to return a single scoped
store for a particular child, from its identifier. This is useful to
avoid re-scoping all children when only a subset of children need to be
created (e.g. in a diffing set up, where only insertions need to be
created).

Finally, a new `DiffableListsOfState` UIKit example was created to
demonstrate the usage of these new primitives in a real life scenario.
The example showcases a setup with a list of (collection view) cells
that have a store and are backed by a diffable data source.

## Changes

- Create new `IdentifiedContainer` protocol, conform `IdentifiedArray`
to it.

- Update `Reducer.forEach` primitive that used `IdentifiedArray` to use
a `IdentifiedContainer` instead.

- Create new `DiffableListsOfState` UIKit example
  + Create new `IdentifiedStateContainer` example data structure that
conforms to `IdentifiedContainer` and wraps a `OrderedDictionary` from
`swift-collections`.
  + Create new DiffableCounterList TCA "stack" and
`DiffableCountersTableViewController` which uses Diffing in a
`UICollectionView` via `UICollectionViewDiffableDataSource`.
  + Create new `CounterCollectionViewCell` which wraps a counter, to
demonstrate injecting child stores in cells.