Loosen platform requirements for Dependencies library (#1466)
* Loosen platform requirements for Dependencies

Dependencies depends on runtime warning functionality, which is brought
in via the "os" module. If we check for availability we can allow the
Dependencies module to be used from multiplatform libraries, as we have
in isowords.

* More agnostic

* wip

* Update RuntimeWarnings.swift

* Update RuntimeWarnings.swift

(cherry picked from commit 2d943dadae3b378f083f9b234d34df1ee62043dd)

# Conflicts:
#	Sources/ComposableArchitecture/Internal/RuntimeWarnings.swift
#	Sources/ComposableArchitecture/SwiftUI/SwitchStore.swift
#	Tests/ComposableArchitectureTests/EffectRunTests.swift
#	Tests/ComposableArchitectureTests/EffectTaskTests.swift