Revert #1802 (#1845)
Fixes #1812.

This regression was subtle, but it's better to retain the old crashing
behavior than introduce this change since it's been with the library for
a very long time, and we have workarounds like `ForEachStore`.

We'll try to re-explore these subtle binding behaviors in the future.

(cherry picked from commit 717796fb4cc05514df633216b8b1c8f4a788a35d)

# Conflicts:
#	Sources/ComposableArchitecture/ViewStore.swift